Fix Windows Warnings

+++ libsrc/protoserver/ProtoClientConnection.cpp
@@ -214,7 +214,7 @@ void ProtoClientConnection::handleNotImplemented()
 void ProtoClientConnection::sendMessage(const google::protobuf::Message &message)
 {
 	std::string serializedReply = message.SerializeAsString();
-	uint32_t size = serializedReply.size();
+	uint32_t size = static_cast<uint32_t>(serializedReply.size());
 	uint8_t sizeData[] = {uint8_t(size >> 24), uint8_t(size >> 16), uint8_t(size >> 8), uint8_t(size)};
 	_socket->write((const char *) sizeData, sizeof(sizeData));
 	_socket->write(serializedReply.data(), serializedReply.length());
